BASEURL Make Client baseUrl empty on default
Base url parameter should not have default value as any new client has
his own `suite uri` like http://suite15.emarsys.net and must use only
that uri. Base uri is part of credentials and it is personal. No common
value for all!

You made it with default value and there is no any warning or error when
first trying to use this package. It's easy to forget to pass right base
uri. It leads to time wasting when trying to understand what's wrong
when API doesn't work but username and secret are correct.